move bp registry
