+++ .prettierignore
@@ -1,1 +1,1 @@
-src/bp-registry/registry.json
+src/core/bp-registry/registry.json
