+++ src/core/bp-registry/1920x1080/th.png
