+++ src/core/bp-registry/1920x1080/zh-CN.png
