+++ src/core/bp-registry/registry.json
