+++ src/core/ocr/ocr.test.ts
@@ -1,7 +1,7 @@
 import { SharpImageContainer } from '@/platform-node';
 import path from 'path';
 import sharp from 'sharp';
-import registry from '../../bp-registry/registry.json';
+import registry from '../bp-registry/registry.json';
 import { BufferSize, DaemonsRawData, GridRawData } from '../common';
 import {
   BreachProtocolFragmentStatus,
